stop unauthorized users from editing and deleting listings
https://youtu.be/MYyJ4PuL4pY?t=15364

+++ app/Http/Controllers/ListingController.php
@@ -63,6 +63,11 @@ class ListingController extends Controller
     public function update(Request $request, Listing $listing)
     {
 
+        // Make sure logged in user is owner
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $formFields = $request->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -85,6 +90,10 @@ class ListingController extends Controller
 
     public function destroy(Listing $listing)
     {
+        if ($listing->user_id != auth()->id()) {
+            abort(403, 'Unauthorized action.');
+        }
+
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
     }
